+++ Sources/Core/Shell/Simctl/Simctl.swift
@@ -54,8 +54,8 @@ public extension Simctl {
         return deviceIDs
     }
 
-    static func killAllSimulators(logInset: Int = 0) {
-        Logger.shared.info("Killing all open simulators", inset: logInset)
+    static func killAllSimulators() {
+        Log.simctl.info("Killing all open simulators")
 
         try? runAndPrint(bash: "killall Simulator")
         try? runAndPrint(bash: "killall iPhone Simulator")
@@ -63,23 +63,23 @@ public extension Simctl {
     }
 
     static func createDevice(name: String, id: String, runtime: Runtime) throws -> String {
-        Logger.shared.info("Create device \(id) with name \"\(name)\" and runtime \(runtime)", inset: 1)
+        Log.simctl.info("Create device \(id) with name \"\(name)\" and runtime \(runtime)")
         return try Simctl._createDevice(name: name, id: id, runtime: runtime)
     }
 
     static func updateStyle(_ style: Style, deviceIds: [String]) throws {
         try deviceIds.forEach {
-            Logger.shared.info("Set style \(style) for device \($0)", inset: 1)
-            try _boot(deviceId: $0, logInset: 1)
+            Log.simctl.info("Set style \(style.rawValue) for device \($0)")
+            try _boot(deviceId: $0)
             try _setAppearance(for: $0, style: style)
         }
     }
 
     static func updateStatusBar(deviceIds: [String]) throws {
         try deviceIds.forEach {
-            Logger.shared.info("Set statusbar for device \($0)", inset: 1)
+            Log.simctl.info("Set statusbar for device \($0)")
 
-            try _boot(deviceId: $0, logInset: 1)
+            try _boot(deviceId: $0)
             try _updateStatusBar(deviceId: $0)
         }
     }
@@ -100,12 +100,12 @@ public extension Simctl {
 
         for style in styles {
             // The following generates a long log of all devices (Useful on a CI for debugging)
-//            Logger.shared.info("Found devices:", inset: 1)
+//            Log.simctl.info("Found devices:", inset: 1)
 //            let devicesForRT = try _list().devices
 //            devicesForRT.keys.forEach({ runtime in
-//                Logger.shared.info("Runtime \(runtime)", inset: 2)
+//                Log.simctl.info("Runtime \(runtime)", inset: 2)
 //                devicesForRT[runtime]!.forEach({ device in
-//                    Logger.shared.info(device, inset: 3)
+//                    Log.simctl.info(device, inset: 3)
 //                })
 //            })
 
@@ -152,16 +152,15 @@ public extension Simctl {
                     throw Error.xcTestRunFileNotFound(path: xcTestRunFile.path())
                 }
 
-                Logger.shared.info("""
-                Running test plan '\(testPlanName) (\(testPlanConfigs.isEmpty ? "all configs" : ListFormatter
-                    .localizedString(byJoining: testPlanConfigs)))' for:
-                    style '\(style)'
+                Log.simctl.info("""
+                Running test plan '\(testPlanName) (\(testPlanConfigs.isEmpty ? "all configs" : ListFormatter.localizedString(byJoining: testPlanConfigs)))' for:
+                    style '\(style.rawValue)'
                     scheme '\(scheme)'
                     runtime: '\(runtime)'
                     platform: '\(platform)'
                     architecture: '\(arch)'
                     xctestrun: '\(xcTestRunFile.path())'
-                """, inset: 1)
+                """)
 
                 // This command just needs the binaries and the path to the
                 // xctestrun file created before the actual testing. Then
@@ -178,9 +177,8 @@ public extension Simctl {
                     derivedDataUrl: derivedDataUrl
                 )
 
-                Logger.shared.info(
-                    "Extracting screenshots from xcresult bundle '\(resultsBundleUrl.path())' for scheme '\(scheme)' and style '\(style)'",
-                    inset: 1
+                Log.simctl.info(
+                    "Extracting screenshots from xcresult bundle '\(resultsBundleUrl.path())' for scheme '\(scheme)' and style '\(style.rawValue)'"
                 )
 
                 try fileManager.createDirectory(at: screensUrl, withIntermediateDirectories: true, attributes: nil)
@@ -189,19 +187,21 @@ public extension Simctl {
         }
 
         for scheme in schemes {
-            Logger.shared.info("Package files into one ZIP for scheme '\(scheme)'", inset: 1)
+            Log.simctl.info("Package files into one ZIP for scheme '\(scheme)'")
 
             let originalDirectoryPath = fileManager.currentDirectoryPath
 
             // Switch into folder to prevent storage of absolute paths
             fileManager.changeCurrentDirectoryPath(outUrl.path)
 
+            defer {
+                // Switch back to the original directory
+                fileManager.changeCurrentDirectoryPath(originalDirectoryPath)
+            }
+
             try Zip.zip(outFile: zipFileName,
                         relativeTargetFolder: scheme,
                         excludePattern: "*.xcresult*")
-
-            // Switch back to the original directory
-            fileManager.changeCurrentDirectoryPath(originalDirectoryPath)
         }
     }
 }
@@ -274,9 +274,9 @@ public extension Simctl {
 
     enum BatteryLevel: Int {
         case empty = 0
-        case quater = 25
+        case quarter = 25
         case fifty = 50
-        case threeQuater = 75
+        case threeQuarter = 75
         case full = 100
     }
 }
@@ -289,7 +289,7 @@ private extension Simctl {
         let out = run(bash: "xcrun simctl list --json")
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
 
@@ -307,14 +307,14 @@ private extension Simctl {
         return id
     }
 
-    static func _boot(deviceId: String, logInset: Int = 0) throws {
-        Logger.shared.info("Boot device \(deviceId)", inset: logInset)
+    static func _boot(deviceId: String) throws {
+        Log.simctl.info("Boot device \(deviceId)")
 
         // Wait while the simulator is booting (https://stackoverflow.com/a/56267933/971329)
         let out = run(bash: "xcrun simctl bootstatus '\(deviceId)' -b")
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
     }
@@ -323,7 +323,7 @@ private extension Simctl {
         let out = run(bash: "xcrun simctl shutdown '\(deviceId)'")
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
     }
@@ -332,7 +332,7 @@ private extension Simctl {
         let out = run(bash: "xcrun simctl ui '\(deviceId)' appearance '\(style.rawValue)'")
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
     }
@@ -347,7 +347,7 @@ private extension Simctl {
         let out = run(bash: "xcrun simctl create '\(name)' '\(id)' '\(runtime.identifier)'")
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
         return out.stdout
@@ -378,7 +378,7 @@ private extension Simctl {
         let out = run("xcrun", args)
 
         if let error = out.error {
-            Logger.shared.error(out.stderror)
+            Log.simctl.error("\(out.stderror)")
             throw error
         }
     }